backend/requirements.txt
Flask
flask-cors
PyPDF2
python-pptx
sentence-transformers
faiss-cpu
numpy
